refactor: move file API to separate file
